Plot a heatmap.

This is a wrapper of the ComplexHeatmap::Heatmap() with additional functions and more friendly interfaces for some common tasks in plotting heatmaps.

Usage

cHeatmap(
  mat1,
  whiteValue = NA,
  colMap = c(green4 = NA, white = whiteValue, red = NA),
  intAsDiscreteCutoff = 6,
  legendPos = c("right", "left", "bottom", "top"),
  nRowCluster = NULL,
  nColmCluster = NULL,
  rowAnnoDf = NULL,
  colmAnnoDf = NULL,
  rowAnnoPara = list(na_col = "white"),
  colmAnnoPara = list(na_col = "white"),
  rowAnnoColMap = list(),
  colmAnnoColMap = list(),
  drawHeatmap = T,
  resetOutliers = is.numeric(mat1),
  clusteringUsingResetValues = FALSE,
  cellFun = NULL,
  cellFontSize = 9,
  cellFontColor = "black",
  rowDraw = NULL,
  ...
)

Arguments

mat1

A numeric or character matrix or data frame, required.

whiteValue

numeric, NULL. If mat1 is numeric, the value of the white color in the legend.

colMap

A vector, c("green4"=NA, "white"=whiteValue, "red"=NA). It defines the colors used in the heatmap and their mapping to values, and therefore should be supplied as a named vector.

For character mat1, colMap is a character vector, i.e. c('red'='a','pink'='b','yellow'='c') if the unique values in mat1 are a, b, and c.

For numeric mat1, colMap is a numeric vector, i.e. c('blue'= NA,'green'=5, 'red'=10)

colMap can be supplied w/o names for convenience, i.e. c(-1, 0, 3); in this case, the length must be three and names are assumed to be c("green4", "white", "red")

the 1st and last value of colMap are the lower and upper bounds of displayed values; values out side of that range will be considered as outliers.

one or both of the 1st and last values of colMap can be supplied as NA; then outliers will be auto-computed if resetOutliers is TRUE.

intAsDiscreteCutoff

integer, 6; if mat1 is an integer matrix and has < 6 unique values, its legend color mapping will be displayed as discrete.

legendPos

one of c("right", "left", "bottom", "top")

nRowCluster, nColmCluster

integer, NULL; number of colored clusters in row or column; different clusters will be in different colors; If this argument is set, it assumes cluster_rows and cluster_columns are TRUE; so do NOT explicitly set those two arguments again. To also split column, set column_split = nColmCluster

rowAnnoDf, colmAnnoDf

matrix or data frame, NULL; annotation for row or column. The rows of the rowAnnoDf/colmAnnoDf should be of the same length and order of the row/column of mat1.

rowAnnoPara, colmAnnoPara

list, list(na_col = "white"); passed to ComplexHeatmap::rowAnnotation() or ComplexHeatmap::columnAnnotation()

rowAnnoColMap, colmAnnoColMap

list, list(); The color mappings for row and column annotations are handled automatically. Use this parameter to manually set the color mappings, names(list) is the names of each annotation. Each item of the list is a vector named by colors, similar to parameter colMap.

If an annotation is numeric, the vector is numeric , i.e. c("green4"= -1, "yellow"=0, "red"=5)

If an annotation is not numeric, the vector is character, i.e. c('red'='a','pink'='b','yellow'='c')

drawHeatmap

logical, TRUE; it uses ComplexHeatmap::draw() to merge all legends into one columns and draw the final heatmap.

It should be set to TRUE to obtain the clustering orders, i.e. by column_order()

However, it should be FALSE if the returned heatmap from cHeatmap() will be concatenated with another heatmaps.

resetOutliers

logical, is.numeric(mat1); outlier values in mat1 can squeeze the whole color scale. If TRUE, outliers will be reset to the max or min values excluding them. The values of those outliers can be displayed directly on the heatmap at cellFontSize and cellFontColor.

clusteringUsingResetValues

logical, FALSE; should the clustering of rows and columns use the reset values of outliers?

cellFun

NULL; it is used to display info for individual cells; text info can be formatted by cellFontSize and cellFontColor. here are different usages:

A function to determine where and what to display at certain cells; it takes a cell value as input and returns a number or character. for example:

cellFun = function(x){x}: display the value of each cell

cellFun = function(x){if(x > 10) x}: display the value of cells whose values are greater than 10

cellFun = function(x){if(x > 10) '?'}: display the '?' sign

cellFun = function(x){if(x > 10) list('rect',col='black',lwd=2)}: color the edges black and set line width as 2. Check the parameters of grid::gpar() for details.

A character vector or list:

cellFun = 'o': display the values of all the outliers in mat1 if any

cellFun = c('o','+'): display the '+' sign instead of the value at the outlier cells

cellFun=c('o',list('rect',col='red',lwd=2)): color the outliers cells with red edges

The 'o' and 'rect' are hard coded cases for usage convenience; for more complicated cell drawing, use the parameter cell_fun of ComplexHeatmap::Heatmap() directly.

cellFontSize

numeric, 9

cellFontColor

character, "white"

rowDraw

list, NULL; it is used to plot across rows. The list should contain the following item:

the function in the grid package to make the draw and its parameters; for examples:

list('grid.points',size = grid::unit(0.3, "char"),gp=grid::gpar(col='black')) ; see grid::grid.points() for details. For grid.point, it can also be simply coded as list('grid.points',size = 0.3,pch=1,col='black') parameter col and following are for gpar().

list('grid.lines',gp=grid::gpar(col='black',lwd = 2)); see grid::grid.lines() for details. For grid.lines, it can also be simply coded as list('grid.lines',col='black',lwd = 2) parameter col and following are for gpar().

list( list('grid.points',size = 0.3,col='black'), list('grid.lines',col='black',lwd = 2) ) for multiple drawings

a matrix containing the data to make the drawing, meeting ncol() == ncol(mat1)

an integer vector specifying the rows to make the drawing; if NULL, it means all the rows of mat1 and assumes nrow() == nrow(mat1) for the supplied matrix at #2

See the parameter layer_fun of ComplexHeatmap::Heatmap() for more complex drawing across a block of cells.

...

passed to ComplexHeatmap::Heatmap()

Value

a ComplexHeatmap::HeatmapList object if drawHeatmap, or a ComplexHeatmap::Heatmap object otherwise.

Details

Here are the additional function:

automatically or manually set the value of outliers in the input matrix so that the color of the heatmap won't be dominated by the outliers.

the option to set the value mapping to each color in the heatmap legend.

auto coloring of the dendrogram

easily highlight or display the values of certain cells

Examples

#none for now.